backend/requirements.txt
python-dotenv
fastapi
uvicorn[standard]
pydantic
openai==1.6.0
groq
httpx
pydub